+++ tests/urls.py
@@ -1,7 +1,7 @@
 from django.conf.urls import url
 
-from .views import (FieldTestAllView, FieldTestView, ManyToManyView, ManyToOneView, OneToOneView, OverrideGetFieldsView,
-                    OverrideGetQuerysetView)
+from .views import (FieldTestAllView, FieldTestView, ManyToManyView, ManyToOneView, OneToOneView,
+                    OverrideGetCSVWriterFmtParamsView, OverrideGetFieldsView, OverrideGetQuerysetView)
 
 urlpatterns = [
     url(r'^fields/$', FieldTestView.as_view(), name='fields'),
@@ -11,4 +11,6 @@ urlpatterns = [
     url(r'^one-to-one/$', OneToOneView.as_view(), name='one-to-one'),
     url(r'^override-get-queryset/$', OverrideGetQuerysetView.as_view(), name='override-get-queryset'),
     url(r'^override-get-fields/$', OverrideGetFieldsView.as_view(), name='override-get-fields'),
+    url(r'^override-get-csv-writer-fmtparams/$', OverrideGetCSVWriterFmtParamsView.as_view(),
+        name='override-get-csv-writer-fmtparams'),
 ]