+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - More documentation in the README.
+- More tests.
 
 ## [0.2.0] - 2017-08-07
 ### Added