+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - More tests.
 - Check that `get_context_data(self, **kwargs)` is not being overridden.
 
+### Fixed
+- Fixed issue with unicode csv data on Python 2.
+
 ## [0.2.0] - 2017-08-07
 ### Added
 - Allow `get_fields(self, queryset)` to be overridden.